requirements.txt
bcrypt==3.1.1
cffi==1.8.3
click==6.6
cryptography==1.5.2
Flask==0.11.1
Flask-Cors==3.0.2
flask-socketio
idna==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pyasn1==0.1.9
pycparser==2.17
PyJWT==1.4.2
six==1.10.0
Werkzeug==0.11.11
pymongo
gevent
gevent-websocket
requests